requirements.txt
streamlit>=1.22.0
langchain>=0.0.267
openai>=0.27.8
python-dotenv>=1.0.0
pandas>=2.0.3
sqlalchemy>=2.0.19
langgraph>=0.0.11
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4